requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
async-lru==1.0.2
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.4
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-markdownx==3.0.1
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
jdatetime==3.6.2
kombu==5.1.0
Markdown==3.3.4
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyaes==1.6.1
Pyrogram==1.2.9
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
TgCrypto==1.2.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
